Rename immutable predicates
