+++ src/internal/__test__/compare-testutil.js
@@ -1,6 +1,6 @@
 // @flow
 import { fromJS, Record } from 'immutable';
-import {isRecord} from '../predicates';
+import {isRecord} from '../immutableJsPredicates';
 
 let defaultValues = {
     a: null,
