+++ src/internal/immutableJsPredicates.js
