+++ src/util/isAssociative.js
@@ -1,5 +1,5 @@
 // @flow
-import {isAssociative, isImmutable} from '../internal/predicates';
+import {isAssociative, isImmutable} from '../internal/immutableJsPredicates';
 import isObject from './isObject';
 
 export default (thing: *): boolean => isImmutable(thing)
