+++ src/util/isCollection.js
@@ -1,5 +1,5 @@
 // @flow
-import {isCollection, isImmutable} from '../internal/predicates';
+import {isCollection, isImmutable} from '../internal/immutableJsPredicates';
 import isObject from './isObject';
 
 export default (thing: *): boolean => isImmutable(thing)
