+++ src/util/isImmutable.js
@@ -1,3 +1,3 @@
 // @flow
-import {isImmutable} from '../internal/predicates';
+import {isImmutable} from '../internal/immutableJsPredicates';
 export default isImmutable;
