+++ src/util/isIndexed.js
@@ -1,4 +1,4 @@
 // @flow
-import {isIndexed} from '../internal/predicates';
+import {isIndexed} from '../internal/immutableJsPredicates';
 
 export default (thing: *): boolean => isIndexed(thing) || Array.isArray(thing);
