+++ src/util/isOrdered.js
@@ -1,4 +1,4 @@
 // @flow
-import {isOrdered} from '../internal/predicates';
+import {isOrdered} from '../internal/immutableJsPredicates';
 
 export default (thing: *): boolean => isOrdered(thing) || Array.isArray(thing);
